Relocated the "model_needs_login()" function to fritzbox_helper.py since it needs to be reused for additional scripts.
Fixed fritzbox_connection_uptime.py by implementing model_needs_login() - logic
