+++ fritzbox_traffic.py
@@ -18,23 +18,7 @@ import os
 import sys
 
 from fritzconnection import FritzConnection
-
-
-def model_needs_login(model):
-    """
-    Checks if provided fritzbox model is in the list of devices which need an authenticated login
-    to extract traffic data.
-
-    :param model: String describing the fritzbox model; as returned from FritzConnection().modelname
-    :return bool
-    """
-    models_require_login = [
-        "FRITZ!Box 6490 Cable",
-    ]
-    for m in models_require_login:
-        if model.startswith(m):
-            return True
-    return False
+from fritzbox_helper import model_needs_login
 
 
 def print_values():
